create working game with result Display!
some overloading of the checkGameResult and calling
it multiple times, due to including too much logic in the function,
caused some serious confusion and frustration making the game not work.

this has now been resolved hopefully for good, with a lesson for next
time to use boolean return values for functions which will determine
a next step (placeMarker) and keeping logic seperate for checks.

for example, creating a function for gameflow which checks if move valid
and only then checkresult etc.